+++ RELEASE_NOTES.md
@@ -5,6 +5,7 @@ Changes:
 * Add consumables to Ultima Runs
 * Add "," to Ultima metrics (1000 -> 1,000) and reduce Total Beads to 6 decimal precision
 * Add "Control Sample Bases > Q30 %" Ultima metric
+* Update Illumina run directory scanning to be completed in reverse order by run directory name (approximately reverse chronological order)
 
 # 2.5.1